Updates for Pottercraft and fixes to Nox (#363)
- fixed knut material
- added new event fired when players try a wand that isn't their destined wand
- changed player wand wood/core accessors so they don't return null
- made wandLoreConjunction public for use elsewhere
- added a get random type for O2WandCoreType and O2WandWoodType
- changed OllivandersAPI to only make objects if they are not initialized... derp
- made Nox also target the caster
- updated description on RemovePotionEffect
- fixed targeted self so it also still targets others for RemovePotionEffectInRadius
- fixed bug where it was not checking wg enabled before trying to set required flags

+++ Ollivanders/src/net/pottercraft/ollivanders2/spell/ET_INTERFICIAM_ANIMAM_LIGAVERIS.java
@@ -62,7 +62,8 @@ public final class ET_INTERFICIAM_ANIMAM_LIGAVERIS extends O2Spell
       initSpell();
 
       // world guard flags
-      worldGuardFlags.add(Flags.MOB_SPAWNING); // needed because Fiendfyre requires it, otherwise horcruxes could get made in locations players couldn't kill them
+      if (Ollivanders2.worldGuardEnabled)
+         worldGuardFlags.add(Flags.MOB_SPAWNING); // needed because Fiendfyre requires it, otherwise horcruxes could get made in locations players couldn't kill them
    }
 
    @Override
